+++ src/io/element.h
@@ -3,9 +3,9 @@
 
 class Element {
 	int m;
-	int value;
+	double value;
 public:
-	Element(int m, int value);
+	Element(int m, double value);
 	double getValue();
 	int getRow();
 	bool operator()(const Element* lhs, const Element* rhs) const  {
